Switch from RuntimeDirectory to systemd-tempfiles
Both redis-server and redis-sentinel use the the same `RuntimeDirectory`
(/run/redis). This is wrong since systemd removes RuntimeDirectory on
service stop. So, stopping redis-server removes redis-sentinel.pid as
well.

Using a systemd-tempfile is a more robust approach. We are also removing
ExecStartPre lines since directory creation is handled in a different
level.